fixed friends styling
