+++ client/components/friends/index.css
@@ -1,0 +1,3 @@
+.card-area {
+  margin-left: 7.5%
+}
